requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.2
werkzeug==3.1.3
jinja2==3.1.6
sqlalchemy==2.0.40
requests==2.32.3
pillow==11.1.0
openai==1.75.0
python-dotenv==1.1.0
schedule==1.2.2
jieba==0.42.1
wtforms==3.2.1
email-validator==2.1.1
gunicorn==21.2.0